+++ tests/unit/commands/test_analyze.py
@@ -97,12 +97,12 @@ class TestAnalyze(object):
                 "Error: Rule format is invalid. Must be in JSON format.\n",
             ),
             (
-                ["analyze", "--rule", "{}", "1"],
-                "Error: Rule format is invalid. Must be in list format.\n",
+                ["analyze", "--rule", '"1"', "1"],
+                "Error: Rule format is invalid. Must be in list or object format.\n",
             ),
             (
-                ["analyze", "1", "--rule", "{}"],
-                "Error: Rule format is invalid. Must be in list format.\n",
+                ["analyze", "1", "--rule", '"1"'],
+                "Error: Rule format is invalid. Must be in list or object format.\n",
             ),
         ],
     )